tests/submission/W3C/domhtml/doc01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dom.html#document.title" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<p>
<a accesskey="g" charset="US-ASCII" coords="0,0,100,100" dir="LTR" href="../images/yellow.png" hreflang="en" id="Anchor" name="Anchor" rel="GLOSSARY" rev="STYLESHEET" shape="rect" tabindex="22" type="image/png">View Submit Button</a>
</p>


<script>
var domtest = async_test("Retrieve the title attribute of HTMLDocument and examine it's value.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var vtitle;
      var doc;
      doc = window.document;
      vtitle = doc.title;

      assert_equals(vtitle,"NIST DOM HTML Test - Anchor");
       
    domtest.done();
}));

</script>
<p>Retrieve the title attribute of HTMLDocument and examine it's value.</p>
<div id="log"></div>
</body></html>
